+++ lib/rubocop/cop/metrics/line_length.rb
@@ -99,7 +99,8 @@ module RuboCop
         end
 
         def allowed_uri_position?(line, uri_range)
-          uri_range.begin < max && uri_range.end == line.length
+          uri_range.begin < max &&
+            (uri_range.end == line.length || uri_range.end == line.length - 1)
         end
 
         def find_excessive_uri_range(line)